Spin off the owner email editor into it's own class.
