+++ lib/pushmi_pullyu.rb
@@ -11,6 +11,7 @@ require 'pushmi_pullyu/aip/downloader'
 require 'pushmi_pullyu/aip/solr_fetcher'
 require 'pushmi_pullyu/aip/fedora_fetcher'
 require 'pushmi_pullyu/aip/file_list_creator'
+require 'pushmi_pullyu/aip/owner_email_editor'
 require 'active_record'
 require 'pushmi_pullyu/aip/user'
 require 'pushmi_pullyu/cli'
