+++ lib/pushmi_pullyu/aip/fedora_fetcher.rb
@@ -2,10 +2,7 @@ require 'net/http'
 
 class PushmiPullyu::AIP::FedoraFetcher
 
-  OWNER_PREDICATE = RDF::URI('http://purl.org/ontology/bibo/owner').freeze
-
   class FedoraFetchError < StandardError; end
-  class NoOwnerPredicate < StandardError; end
 
   RDF_FORMAT = 'text/rdf+n3'.freeze
 
@@ -39,7 +36,7 @@ class PushmiPullyu::AIP::FedoraFetcher
 
     if response.is_a?(Net::HTTPSuccess)
       body = if should_add_user_email
-               add_user_email(response.body)
+               PushmiPullyu::AIP::OwnerEmailEditor.new(response.body).run
              else
                response.body
              end
@@ -63,53 +60,6 @@ class PushmiPullyu::AIP::FedoraFetcher
 
   def base_path
     PushmiPullyu.options[:fedora][:base_path]
-  end
-
-  def add_user_email(body)
-    ensure_database_connection
-
-    is_modified = false
-    prefixes = nil
-    # Read once to load prefixes (the @things at the top of an n3 file)
-    RDF::N3::Reader.new(input = body) do |reader|
-      reader.each_statement { |_statement| }
-      prefixes = reader.prefixes
-    end
-    new_body = RDF::N3::Writer.buffer(prefixes: prefixes) do |writer|
-      RDF::N3::Reader.new(input = body) do |reader|
-        reader.each_statement do |statement|
-          if statement.predicate == OWNER_PREDICATE
-            user = PushmiPullyu::AIP::User.find(statement.object.to_i)
-            writer << [statement.subject, statement.predicate, user.email]
-            is_modified = true
-          else
-            writer << statement
-          end
-        end
-      end
-    end
-    return new_body if is_modified
-    raise NoOwnerPredicate
-  end
-
-  def ensure_database_connection
-    return if ActiveRecord::Base.connected?
-    ActiveRecord::Base.establish_connection(database_configuration)
-  end
-
-  def database_configuration
-    # Config either from URL, or with more granular options (the later taking precedence)
-    config = {}
-    uri = URI.parse(PushmiPullyu.options[:database][:url])
-    config[:adapter] = PushmiPullyu.options[:database][:adaptor] || uri.scheme
-    config[:host] = PushmiPullyu.options[:database][:host] || uri.host
-    config[:database] = PushmiPullyu.options[:database][:database] || uri.path.split('/')[1].to_s
-    config[:username] = PushmiPullyu.options[:database][:username] || uri.user
-    config[:password] = PushmiPullyu.options[:database][:password] || uri.password
-    params = CGI.parse(uri.query || '')
-    config[:encoding] = PushmiPullyu.options[:database][:encoding] || params['encoding'].to_a.first
-    config[:pool] = PushmiPullyu.options[:database][:pool] || params['pool'].to_a.first
-    config
   end
 
 end
